backend/requirements.txt
asgiref==3.11.0
asttokens==3.0.1
colorama==0.4.6
decorator==5.2.1
Django==6.0
djangorestframework==3.16.1
executing==2.2.1
ipdb==0.13.13
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
parso==0.8.5
prompt_toolkit==3.0.52
pure_eval==0.2.3
Pygments==2.19.2
sqlparse==0.5.4
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.3
wcwidth==0.2.14
